Fix check subscription
